Rename React.js entrypoint to ReactClient.js
The server entrypoint no longer depends on this module. It's now only
imported by the client. I've updated the name to reflect that it's
client-only, and can include client-only features.
